UPSTREAM: media: venus: venc: Fix default capture format
The default codec after driver open is set to be H264 but the
instance format for capture is wrongly set to H263. Correct this
to H264.

For regular applications this is not a big issue because they set
the format through S_FMT but for example v4l2-compliance does not.

Signed-off-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 4568ea3fa95d33cc10e8938949ae9a875046e553
 https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git master)

BUG=183962425
TEST=emerge-herobrine sys-kernel/chromeos-kernel-5_10

Change-Id: I3415e8b772d9e015cb488bfccb34ef22703e04ca
Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2796843
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Tested-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>