UPSTREAM: ACPI: Execute platform _OSC also with query bit clear
The platform _OSC can change the hardware state when query bit is not
set. According to ACPI spec it is recommended that the OS runs _OSC with
query bit set until the platform does not mask any of the capabilities.
Then it should run it with query bit clear in order to actually commit
the changes. Linux has not been doing this for the reasons that there
has not been anything to commit, until now.

The ACPI 6.4 introduced _OSC for USB4 to allow the OS to negotiate
native control over USB4 tunneling. The platform might implement this so
that it only activates the software connection manager path when the OS
calls the _OSC with the query bit clear. Otherwise it may default to the
firmware connection manager, for instance.

For this reason modify the _OSC support so that we first execute it with
query bit set, then use the returned value as base of the features we
want to control and run the _OSC again with query bit clear. This also
follows what Windows is doing.

Also rename the function to better match what it does.

Signed-off-by: Mario Limonciello <mario.limonciello@dell.com>
Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 719e1f561afbe020ed175825a9bd25ed62ed1697)

BUG=b:186450411
TEST=Build, boot and confirm thunderbolt driver is loaded on brya

Change-Id: Ied2da0f1554266ed86cbe14473fdda7a1b3bf98f
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2852779
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>