UPSTREAM: platform/x86: Intel PMT Crashlog capability driver
Add support for the Intel Platform Monitoring Technology crashlog
interface. This interface provides a few sysfs values to allow for
controlling the crashlog telemetry interface as well as a character
driver to allow for mapping the crashlog memory region so that it can be
accessed after a crashlog has been recorded.

This driver is meant to only support the server version of the crashlog
which is identified as crash_type 1 with a version of zero. Currently no
other types are supported.

Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 5ef9998c96b0c99c49c202054586967e609286d2)

BUG=b:187383885
TEST=Verify kernel builds for brya.
     `emerge-brya chromeos-kernel-5_10`
TEST=Verify PMT sysfs exists on brya.
     `ls /sys/class/intel_pmt`

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: I90f8d5b90e72cfd4286b7cf603443a52c76ef0e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2857081
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>
Tested-by: Azhar Shaikh <azhar.shaikh@intel.com>