UPSTREAM: media: uapi: add MEDIA_BUS_FMT_METADATA_FIXED media bus format.
MEDIA_BUS_FMT_METADATA_FIXED should be used when
the same driver handles both sides of the link and
the bus format is a fixed metadata format that is
not configurable from userspace.
The width and height will be set to 0 for this format.

Signed-off-by: Dafna Hirschfeld <dafna.hirschfeld@collabora.com>
Acked-by: Helen Koike <helen.koike@collabora.com>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 6ad253cc3436269fc6bcff03d704c672f368da0a)

BUG=b:197575417
TEST=Camera works on Scarlet

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Iedfaab944695aa29ed61894713030a6e6e90b7f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3168470
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>