UPSTREAM: thunderbolt: Add USB4 router operation proxy for firmware connection manager
Intel Maple Ridge and Tiger Lake connection manager firmware implements
a USB4 router operation proxy that should be used instead of direct
register access to avoid races with the firmware. This is supported in
all firmwares where the protocol version field returned in the driver
ready response is 3 (or higher).

This adds the USB4 router proxy operations support to the driver so that
we first check the protocol version and if it is 3 (or higher) the USB4
router operation is run through the firmware provided proxy. Otherwise
the native version is used.

Most USB4 router proxy operations are pretty straightforward except
NVM_AUTH where the firmware only responds once the router is restarted
but before it sends device connected notification. To support this we
split the operation so that the reply is received asynchronously and
stored to struct icm. This last reply is then returned in
icm_usb4_switch_nvm_authenticate_status() if available.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 9039387e166edab35c89ddcc057529e332cc4089)

BUG=b:160572503
TEST=No regressions with USB4 usages:
- Hotplug (USB, DP, TBT)
- NVM upgrade
- Suspend

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I5ddd8ffbbb7278e6e178b88ae08284be221044cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2624349
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>