FROMGIT: net: ipa: pass a value to gsi_irq_type_update()
Now that all of the GSI interrupts are handled uniformly,
change gsi_irq_type_update() so it takes a value.  Have the
function assign that value to the cached mask of enabled GSI
IRQ types before writing it to hardware.

Note that gsi_irq_teardown() will only be called after
gsi_irq_disable(), so it's not necessary for the former
to disable all IRQ types.  Get rid of that.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 8194be79fbbc36dae391b7d747ba58459687792a
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:174161909
TEST=Booted trogdor, modem call functional, s/r work

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: Id695f18f8a466290789305b2e4546b6b131ddc5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2559822
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>