FROMGIT: net: ipa: change a warning to debug
When we determine from hardware what the size of IPA memory is
we compare it against what we learned about it from DT.

If DT defines a region that's larger than actual memory, we use the
smaller actual size and issue a warning.

If DT defines a smaller region than actual memory we issue a warning
too.  But in this case the difference is harmless; so rather than
issuing a warning, just provide a debug message instead.

Reorder these checks so the one that matters more is done first.

Reported-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 2c642c48b3620d85f4e0ac66de01195a6096b333
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:174161909
TEST=Booted trogdor, modem call functional, s/r work

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I0c3cb15cbf9b8b85bb4cbdab0ae438b9acab5e56
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2559825
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>