UPSTREAM: thunderbolt: Perform USB4 router NVM upgrade in two phases
The currect code expects that the router returns back the status of the
NVM authentication immediately. When tested against a real USB4 device
what happens is that the router is reset and only after that the result
is updated in the ROUTER_CS_26 register status field. This also seems to
align better what the spec suggests.

For this reason do the same what we already do with the Thunderbolt 3
devices and perform the NVM upgrade in two phases. First start the
NVM_AUTH router operation and once the router is added back after the
reset read the status in ROUTER_CS_26 and expose it to the userspace
accordingly.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 661b19473bf3ac0924560f0cbf84c15458b3c8de)

BUG=b:160572503
TEST=No regressions with USB4 usages:
- Hotplug (USB, DP, TBT)
- NVM upgrade
- Suspend

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I5fe93e38c2e0a84e9f736ee90560edf1d8bfa754
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2623644
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>