misc: cs40l25: Rewritable GPIO virtual waveform support
Adds the ability to create and trigger a virtual waveform which
can be assigned to GPIO1 rise or fall index.  This effect is not
saved but can be overwritten as many times as necessary.  This is
done by setting a special cp_trigger_index (65527 or 65528)
immediately after sending a new cp_trigger_queue or pwle effect.
Example for gpio1_fall_index:
echo 65527 > path/to/device/cp_trigger_index
Example for gpio1_rise_index:
echo 65528 > path/to/device/cp_trigger_index

Change-Id: Ida42e68d796dccf76e439f51f2977f41b8dc5aed
Signed-off-by: Tim Howe <tim.howe@cirrus.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>