FROMGIT: remoteproc/mediatek: skip if filesz is 0
The main purpose of the loop is to load the memory to the SCP SRAM.
If filesz is 0, can go to next program header directly.

We don't need to try to validate the FW binary for those filesz==0
segments.

Reviewed-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Signed-off-by: Tzung-Bi Shih <tzungbi@google.com>
Link: https://lore.kernel.org/r/20201116084413.3312631-3-tzungbi@google.com
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 48cb5b6829e2acb4b600153142889f879115e172
 git://git.kernel.org/pub/scm/linux/kernel/git/andersson/remoteproc.git for-next)

BUG=b:173086286
TEST=emerge-asurada chromeos-kernel-5_4 && boot SCP

Signed-off-by: Tzung-Bi Shih <tzungbi@chromium.org>
Change-Id: I21b897f160ce5a3d599d5c8ede57ddefc1d0ece0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2534432