FROMLIST: sched: Core-wide rq->lock
Introduce the basic infrastructure to have a core wide rq->lock.

This relies on the rq->__lock order being in increasing CPU number
(inside a core). It is also constrained to SMT8 per lockdep (and
SMT256 per preempt_count).

Luckily SMT8 is the max supported SMT count for Linux (Mips, Sparc and
Power are known to have this).

BUG=b:152605392
TEST=run power_VideoCall test

(am from
https://lore.kernel.org/lkml/20210422123308.256677625@infradead.org/)

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Don Hiatt <dhiatt@digitalocean.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: Ib26f4510ff9de18147f86b4e216c7abfc3c87831
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2880780
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>