misc: cs40l25: Fix system suspend IRQ handling
Runtime PM is disabled during system suspend, and on system resume
IRQs are enabled before the runtime PM is. As the cs40l25's IRQ
handler does a PM runtime get this will lead to issues if an IRQ
is handling in that window.

To handle this issue we disable the IRQ temporarily during the
suspend and resume processes, re-enabling it in the late stages of
each. This means the IRQ is still active during system suspend but
can not be handled during the problematic windows.

Change-Id: Ic31944f9e6a494491d274c1fd5981163b5ab1b6c
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>