input: cs40l26: Updates to power on sequence
The power on sequence has 32 value-address pairs with a
total of 64 registers to populate.
Ensure that cs40l26_pseq_init() parses all possible registers.

Add option to replace or not replace an address-value pair
if address is already present in the power on sequence.
This will accommodate any addresses that require multiple
writes after exiting hibernation.

Change-Id: I2e78e1625c4b4f18a5068f07233c16e1890870db
Signed-off-by: Fred Treven <fred.treven@cirrus.com>