FROMGIT: remoteproc: sysmon: Expose the shutdown result
A graceful shutdown of the Qualcomm remote processors where
traditionally performed by invoking a shared memory state signal and
waiting for the associated ack.

This was later superseded by the "sysmon" mechanism, where some form of
shared memory bus is used to send a "graceful shutdown request" message
and one of more signals comes back to indicate its success.

But when this newer mechanism is in effect the firmware is shut down by
the time the older mechanism, implemented in the remoteproc drivers,
attempts to perform a graceful shutdown - and as such it will never
receive an ack back.

This patch therefor track the success of the latest shutdown attempt in
sysmon and exposes a new function in the API that the remoteproc driver
can use to query the success and the necessity of invoking the older
mechanism.

Tested-by: Steev Klimaszewski <steev@kali.org>
Reviewed-by: Rishabh Bhatnagar <rishabhb@codeaurora.org>
Link: https://lore.kernel.org/r/20201122054135.802935-3-bjorn.andersson@linaro.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 5c212aaf5457ca5bd99aba3ad29a4a17f8129939
 https://git.kernel.org/pub/scm/linux/kernel/git/andersson/remoteproc.git/ rproc-next)

BUG=b:149924251
TEST=Boot trogdor modem

Signed-off-by: Evan Green <evgreen@chromium.org>
Change-Id: I51efbec23a31b47c6883a8de68ecc66e7e247730
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2566438
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Sujit Kautkar <sujitka@chromium.org>