FROMGIT: drivers: qcom: rpmh-rsc: Do not read back the register write on trigger
When triggering a TCS to send its contents, reading back the trigger
value may return an incorrect value. That is because, writing the
trigger may raise an interrupt which could be handled immediately and
the trigger value could be reset in the interrupt handler.

A write_tcs_reg_sync() would read back the value that is written and try
to match it to the value written to ensure that the value is written,
but if that value is different, we may see false error for same.

Reviewed-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Signed-off-by: Maulik Shah <mkshah@codeaurora.org>
Link: https://lore.kernel.org/r/1606211610-15168-1-git-send-email-mkshah@codeaurora.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit e1d8008179fef782ecea4e7af1b9cd9891bd881e
 git://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git drivers-for-5.12)

BUG=b:176902912
TEST=Nothing changes

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I420ca6eeee0b77102b0dffe8d6af213633210221
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2613305
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>