UPSTREAM: i2c: remove check that can never be true
A driver remove callback is only called if the device was bound before.
So it's sure that both dev and dev->driver are valid and dev is an i2c
device. If the check fails something louder than "return 0" might be
appropriate because the problem is grave (something like memory
corruption), otherwise the check is useless.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Wolfram Sang <wsa@kernel.org>

(cherry picked from commit 4e970a0ada5299d017a4263074f725227c2d2852)

BUG=none
TEST=boot krane

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I13f9caa4aa98df41c670ae3f33615edfa1563da5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2873328
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>