CHROMIUM: iwl7000: mvm: fix the return type for DSM functions 1 and 2
The return type value of functions 1 and 2 were considered to be an
integer inside a buffer, but they are only an integer, without the
buffer.  Fix the code in iwl_acpi_get_dsm_u8() to handle it as a single
integer value, without expecting it to be packed inside a buffer.

Change-Id: Ic24524627b899c9a01af38107a62a626bdf5ae3a
Signed-off-by: Matt Chen <matt.chen@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 09bf59fbd4b944d23d0e2e5c4511191fe92bf6ba