UPSTREAM: media: controls: Log H264 stateless controls in .std_log
For now, only print the type of the control. In the future,
we might want to be able to print the values of all
the stateless codec controls currently set in the control handler,
which should be useful for debugging reasons.

Signed-off-by: Ezequiel Garcia <ezequiel@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>

(cherry picked from commit afc5329ea76d0706d0db6d26ff98b890bba1d112)

BUG=b:187685837
TEST=boot krane

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I498dc0a031856c2c1e39a8b7f9dcab0b408c7d3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2887126
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>