FROMGIT: BACKPORT: drm/i915/gt: Move the breadcrumb to the signaler if completed upon cancel
If while we are cancelling the breadcrumb signaling, we find that the
request is already completed, move it to the irq signaler and let it be
signaled.

v2: Tweak reference counting so that we only acquire a new reference on
adding to a signal list, as opposed to a hidden i915_request_put of the
caller's reference.

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Tvrtko Ursulin <tvrtko.ursulin@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20201126140407.31952-5-chris@chris-wilson.co.uk
(cherry picked from commit 85cc2917a3965a3a747a6407d6e3028cfeb1534e
https://anongit.freedesktop.org/git/drm-intel.git drm-intel-next)

BUG=b:179453336
TEST=eDP comes up on brya when built with this commit https://crrev.com/i/3584703

Resolved Conflicts:
	drivers/gpu/drm/i915/gt/intel_breadcrumbs.c
due to commit "142c6a6040de drm/i915: Check for rq->hwsp validity after
acquiring RCU lock" already present on chromeos-5.10

Change-Id: Ic7244818238d8baafea7fcefe1579baae862c266
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>