UPSTREAM: media: rockchip: rkisp1: remove useless debugfs checks
The debugfs_create_dir() function never returns NULLs so this code will
never be executed.  It's not intended that callers will check for
debugfs errors in the normal case and it's not necessary in this driver,
so we can just delete this code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Acked-by: Helen Koike <helen.koike@collabora.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab+huawei@kernel.org>
(cherry picked from commit 42ad70c469665387e8f7b53cb6d4581492dff906)

BUG=b:197575417
TEST=Camera works on Scarlet

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Id09ce9a70aa6d5164c7c8e95b212407a90fac38e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3168477
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Ricardo Ribalda <ribalda@chromium.org>