FROMLIST: iio: sx9310: Fix semtech,avg-pos-strength setting when > 16
This DT property can be 0, 16, and then 64, but not 32. The math here
doesn't recognize this slight bump in the power of 2 numbers and
translates a DT property of 64 into the register value '3' when it
really should be '2'. Fix it by subtracting one more if the number being
translated is larger than 31. Also use clamp() because we're here.

Cc: Daniel Campello <campello@chromium.org>
Cc: Lars-Peter Clausen <lars@metafoo.de>
Cc: Peter Meerwald-Stadler <pmeerw@pmeerw.net>
Cc: Douglas Anderson <dianders@chromium.org>
Cc: Gwendal Grignou <gwendal@chromium.org>
Cc: Evan Green <evgreen@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>
(am from https://lore.kernel.org/patchwork/patch/1348464/)
(also found at https://lore.kernel.org/r/20201202200252.986230-1-swboyd@chromium.org)

BUG=b:166572529
TEST=`grep 17: /sys/kernel/debug/regmap/5-0028/registers` shows 0x0a

Change-Id: I1dcd90536ed3b68f2ec8ca260559b5692bb47bb5
Signed-off-by: Guenter Roeck <groeck@chromium.org>